refactor(i18n): consolidate locale files
Keep web/i18n/locales as the single runtime source and preserve keys that existed only in the duplicate directories.

Refs #1019
