+++ web/i18n/locales/ja.json
@@ -164,5 +164,24 @@
       }
     },
     "disclaimer": "Nightbot、StreamElements、Moobot、Fossabot は各社の商標です。Twir はこれらと提携・承認されているわけではありません。機能データは公開情報に基づいており、変更される場合があります。"
+  },
+  "sidebar": {
+    "quotes": "引用"
+  },
+  "quotes": {
+    "title": "引用",
+    "create": "引用を作成",
+    "edit": "引用を編集",
+    "text": "引用",
+    "number": "番号",
+    "creator": "作成者",
+    "game": "ゲーム",
+    "createdAt": "作成日",
+    "deleteConfirmation": "この引用を削除してもよろしいですか？",
+    "created": "引用を作成しました",
+    "updated": "引用を更新しました",
+    "deleted": "引用を削除しました",
+    "errorOnSave": "引用を保存できませんでした",
+    "errorOnDelete": "引用を削除できませんでした"
   }
 }
