+++ web/i18n/locales/ru.json
@@ -61,7 +61,8 @@
     "create": "Создать",
     "edit": "Редактировать",
     "noData": "Нет данных",
-    "loading": "Загрузка..."
+    "loading": "Загрузка...",
+    "show": "Показать"
   },
   "sidebar": {
     "dashboard": "Панель управления",
@@ -94,7 +95,8 @@
     "darkTheme": "Тёмная тема",
     "lightTheme": "Светлая тема",
     "notifications": "Уведомление",
-    "giveaways": "Розыгрыши"
+    "giveaways": "Розыгрыши",
+    "quotes": "Цитаты"
   },
   "mcpGuide": {
     "title": "Доступ для ИИ",
@@ -1731,5 +1733,21 @@
       }
     },
     "disclaimer": "Nightbot, StreamElements, Moobot и Fossabot — товарные знаки их соответствующих владельцев. Twir не аффилирован с ними и не имеет их одобрения. Данные о функциях основаны на публичной информации и могут меняться со временем."
+  },
+  "quotes": {
+    "title": "Цитаты",
+    "create": "Создать цитату",
+    "edit": "Редактировать цитату",
+    "text": "Цитата",
+    "number": "Номер",
+    "creator": "Автор",
+    "game": "Игра",
+    "createdAt": "Создано",
+    "deleteConfirmation": "Вы уверены, что хотите удалить эту цитату?",
+    "created": "Цитата создана",
+    "updated": "Цитата обновлена",
+    "deleted": "Цитата удалена",
+    "errorOnSave": "Не удалось сохранить цитату",
+    "errorOnDelete": "Не удалось удалить цитату"
   }
 }
