+++ web/i18n/locales/uk.json
@@ -94,7 +94,8 @@
     "darkTheme": "Темна тема",
     "lightTheme": "Світла тема",
     "notifications": "Сповіщення",
-    "giveaways": "Розіграші"
+    "giveaways": "Розіграші",
+    "quotes": "Цитати"
   },
   "deleteConfirmation": {
     "text": "Ви впевнені, що хочете видалити це?",
@@ -744,7 +745,9 @@
         "minLength": "Мінімальна тривалість пісні для замовлення (у хвилинах)",
         "maxLength": "Максимальна тривалість пісні для замовлення (у хвилинах)",
         "minViews": "Мінімальна кількість переглядів пісні для замовлення"
-      }
+      },
+      "hideOnPause": "Ховати оверлей при паузі",
+      "hideOnPauseDescription": "Коли увімкнено, OBS оверлей буде ховатися при паузі відтворення. Коли вимкнено, оверлей залишається видимим зі станом паузи."
     },
     "tabs": {
       "general": "Загальні",
@@ -757,6 +760,15 @@
       "song": "Заборонити пісню",
       "userConfirm": "Ви впевнені, що хочете заблокувати цього користувача?",
       "songConfirm": "Ви впевнені, що хочете заблокувати цю пісню?"
+    },
+    "links": {
+      "title": "Посилання на віджет та оверлей",
+      "apiKey": "API ключ",
+      "widget": "Віджет",
+      "overlay": "OBS оверлей",
+      "notConfigured": "Налаштуйте замовлення пісень, щоб отримати посилання на віджет та оверлей.",
+      "copied": "Посилання на {label} скопійовано!",
+      "copyError": "Не вдалося скопіювати посилання"
     }
   },
   "timers": {
@@ -1522,5 +1534,21 @@
       }
     },
     "disclaimer": "Nightbot, StreamElements, Moobot та Fossabot — торгові марки їхніх відповідних власників. Twir не афілійований з ними та не має їхнього схвалення. Дані про функції базуються на публічній інформації та можуть змінюватися з часом."
+  },
+  "quotes": {
+    "title": "Цитати",
+    "create": "Створити цитату",
+    "edit": "Редагувати цитату",
+    "text": "Цитата",
+    "number": "Номер",
+    "creator": "Автор",
+    "game": "Гра",
+    "createdAt": "Створено",
+    "deleteConfirmation": "Ви впевнені, що хочете видалити цю цитату?",
+    "created": "Цитату створено",
+    "updated": "Цитату оновлено",
+    "deleted": "Цитату видалено",
+    "errorOnSave": "Не вдалося зберегти цитату",
+    "errorOnDelete": "Не вдалося видалити цитату"
   }
 }
