+++ web/layers/dashboard/layout/header/layouts/stats-layout-rows.spec.ts
@@ -2,14 +2,14 @@ import { mount } from '@vue/test-utils'
 import { afterEach, describe, expect, it } from 'vitest'
 import { createI18n } from 'vue-i18n'
 
-import de from '../../../../../locales/de.json'
-import en from '../../../../../locales/en.json'
-import es from '../../../../../locales/es.json'
-import ja from '../../../../../locales/ja.json'
-import pt from '../../../../../locales/pt.json'
-import ru from '../../../../../locales/ru.json'
-import sk from '../../../../../locales/sk.json'
-import uk from '../../../../../locales/uk.json'
+import de from '../../../../../i18n/locales/de.json'
+import en from '../../../../../i18n/locales/en.json'
+import es from '../../../../../i18n/locales/es.json'
+import ja from '../../../../../i18n/locales/ja.json'
+import pt from '../../../../../i18n/locales/pt.json'
+import ru from '../../../../../i18n/locales/ru.json'
+import sk from '../../../../../i18n/locales/sk.json'
+import uk from '../../../../../i18n/locales/uk.json'
 import { Platform } from '~/gql/graphql.js'
 import StatsLayoutRows from './stats-layout-rows.vue'
