--- web/locales/ja.json
@@ -1,44 +1,0 @@
-{
-    "languageName": "日本語",
-    "sidebar": {
-        "quotes": "引用"
-    },
-    "dashboard": {
-        "statsLayouts": {
-            "label": "ヘッダー統計レイアウト",
-            "rows": "行",
-            "rowsDescription": "各ウィジェット内にプラットフォーム別の値を表示",
-            "aggregate": "集計",
-            "aggregateDescription": "合計値とプラットフォーム別内訳のポップオーバー",
-            "grid": "グリッド",
-            "gridDescription": "プラットフォームカードのみ、グローバルウィジェットなし"
-        },
-        "platformStats": {
-            "platform": "プラットフォーム",
-            "streamTitles": "配信タイトル",
-            "livePlatforms": "{live} ライブ · {total} プラットフォーム",
-            "breakdown": "プラットフォーム別内訳",
-            "live": "LIVE",
-            "offline": "オフライン",
-            "noTitle": "タイトルなし",
-            "noCategory": "カテゴリなし",
-            "editUnavailable": "編集できません"
-        }
-    },
-    "quotes": {
-        "title": "引用",
-        "create": "引用を作成",
-        "edit": "引用を編集",
-        "text": "引用",
-        "number": "番号",
-        "creator": "作成者",
-        "game": "ゲーム",
-        "createdAt": "作成日",
-        "deleteConfirmation": "この引用を削除してもよろしいですか？",
-        "created": "引用を作成しました",
-        "updated": "引用を更新しました",
-        "deleted": "引用を削除しました",
-        "errorOnSave": "引用を保存できませんでした",
-        "errorOnDelete": "引用を削除できませんでした"
-    }
-}
